+++ screenshots-and-reflections/project-1/change-log.md
@@ -1,3 +1,9 @@
 # Change log
 
 A place for you to document the changes made between code review and final approval submission.
+
+ * Constants and input argument 
+ with output/input paths are shorter
+ * own method for checking right number of cmd arguments
+ * text file is read once
+